[frontend] code improvements
